+++ controllers/userrController.js
@@ -1,0 +1,44 @@
+//Route Handlers
+
+/////////////--------------------USERS---------------------/////////////
+//to get all tours
+exports.get_all_users = (request, response) => {
+  response.status(500).json({
+    status: "error",
+    message: "Method yet to be implemented",
+  });
+};
+
+//to get specified tour
+exports.get_specified_user = (request, response) => {
+  response.status(500).json({
+    status: "error",
+    message: "Method yet to be implemented",
+  });
+};
+
+//to create a new tour
+exports.create_user = (request, response) => {
+  response.status(500).json({
+    status: "error",
+    message: "Method yet to be implemented",
+  });
+};
+
+//to update specified tour
+// to be done after MongoDB is Insitialised
+exports.update_user = (request, response) => {
+  response.status(500).json({
+    status: "error",
+    message: "Method yet to be implemented",
+  });
+};
+
+//to delete a tour
+// to be done after MongoDB is Insitialised
+exports.delete_user = (request, response) => {
+  response.status(500).json({
+    status: "error",
+    message: "Method yet to be implemented",
+  });
+};
